requirements.txt
fastapi
uvicorn
jinja2
azure-identity
azure-mgmt-resource
sqlalchemy
psycopg2-binary